Refactor location of blacklist method
